requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-dotenv==1.0.0
google-generativeai==0.3.2
google-cloud-firestore==2.13.1
aiohttp==3.9.1
pydantic==2.5.0
python-multipart==0.0.6